Rename submission-excel2xml to excel2xml
