+++ lib/excel2xml/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Excel2xml
+  VERSION = '0.1.0'
+end
