--- lib/submission/excel2xml/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Submission
-  module Excel2xml
-    VERSION = "0.1.0"
-  end
-end
