--- sig/submission/excel2xml.rbs
@@ -1,6 +1,0 @@
-module Submission
-  module Excel2xml
-    VERSION: String
-    # See the writing guide of rbs: https://github.com/ruby/rbs#guides
-  end
-end
