+++ spec/excel2xml_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-RSpec.describe Submission::Excel2xml do
+RSpec.describe Excel2xml do
   it "has a version number" do
-    expect(Submission::Excel2xml::VERSION).not_to be nil
+    expect(Excel2xml::VERSION).not_to be nil
   end
 
   it "does something useful" do
