+++ spec/spec_helper.rb
@@ -1,10 +1,10 @@
 # frozen_string_literal: true
 
-require "submission/excel2xml"
+require 'excel2xml'
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
-  config.example_status_persistence_file_path = ".rspec_status"
+  config.example_status_persistence_file_path = '.rspec_status'
 
   # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
